feat: support RDS Enhanced Monitoring
Adds support for RDS Enhanced Monitoring, which shows more detailed
metrics for RDS instances. By default, we disable this for test/staging
stages and set to 60s intervals for production but as always it will be
up to the service owner to decide what they need when they create a
new service.

Jira-Ticket: [AV-24569](https://avrios.atlassian.net/browse/AV-24569)