feat: tag the repository on each stable release
Pushes a tag to the Git repository on each merge to main. This makes
it easier to track version published version changes to the Github
repository.

We talked about having an integrated function within the build environment
itself, stored as a bash function to do this for us but this makes it
more difficult if we have to change the environment (custom runtimes/etc).
Therefore it is easier to solve the problem we have now, by simply using
a template, `jq`, and `curl`.

Jira-Ticket: [AV-24528](https://avrios.atlassian.net/browse/AV-24528)